feat: Calculate single pair of number
- This change ensures that the calculator always calculates a pair of
values. If there are already 2 numbers and 1 operator, the next operator
button will give the answer of that pair of numbers.
- Inside the forEach loop of the operator (operatorList.forEach) add a
variable to store the operator just clicked, if there is already the
first number and an operator, it will give the answer of the
calculation.
- Inside the forEach loop of the number (keylist.forEach) add an
algorithm to solve the case where the screen displays the answer of the
previous calculation and we have to clear it before entering the next
number for the next calculation.